Servlet version 1.5.0b -- the beginning of the next generation :-) - code restructuring to improve scaleability - new Initialiser servlet that must be run first - image transformation work moved to DigilibImageWorker class - Maximum number of concurrent threads limited by Semaphore - old JIMI toolkit implementation removed

+++ servlet/src/digilib/io/ImageFile.java
@@ -129,7 +129,7 @@ public class ImageFile {
 	 * 
 	 * @return
 	 */
-	public double getAspect() {
+	public float getAspect() {
 		return (pixelSize != null) ? pixelSize.getAspect() : 0;
 	}
